requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.17
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==8.4.0
ordered-set==4.1.0
packaging==24.2
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
pyotp==2.9.0
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==2.3.8
wrapt==1.17.2
WTForms==3.2.1
